Reorganize files to have all events creation in one folder
