+++ fe2-android/app/src/androidTest/java/com/github/dedis/student20_pop/detail/fragments/events/creation/MeetingEventCreationFragmentTest.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui.event.creation;
+package com.github.dedis.student20_pop.detail.fragments.events.creation;
 
 public class MeetingEventCreationFragmentTest {
   //  private final int YEAR = 2022;
