+++ fe2-android/app/src/androidTest/java/com/github/dedis/student20_pop/detail/fragments/events/creation/RollCallEventCreationFragmentTest.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui.event.creation;
+package com.github.dedis.student20_pop.detail.fragments.events.creation;
 
 public class RollCallEventCreationFragmentTest {
   //  private final int YEAR = 2022;
