+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/fragments/events/creation/MeetingEventCreationFragment.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui.event.creation;
+package com.github.dedis.student20_pop.detail.fragments.events.creation;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -13,7 +13,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.FragmentManager;
 import com.github.dedis.student20_pop.R;
-import com.github.dedis.student20_pop.detail.fragments.AbstractEventCreationFragment;
 import com.github.dedis.student20_pop.detail.listeners.OnEventCreatedListener;
 
 /**
