+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/fragments/events/package-info.java
@@ -3,4 +3,4 @@
  *
  * <p>The classes inside this package are fragments.
  */
-package com.github.dedis.student20_pop.ui.event;
+package com.github.dedis.student20_pop.detail.fragments.events;
